pyproject.toml
[project]
name = "clio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.0.24",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=3.4.1",
]

[project.optional-dependencies]
transformers = [
    "skrub>=0.5.1",
]

[dependency-groups]
dev = [
    "coverage>=7.6.12",
    "dirty-equals>=0.9.0",
    "jupyter>=1.1.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.5",
    "mkdocstrings-python>=1.16.1",
    "pytest>=8.3.4",
    "ruff>=0.9.7",
]
